+++ src/extension.ts
@@ -215,7 +215,7 @@ export function registerExtension(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.cookCmdAliasDumpWithOthersToFiles',
 		(textEditor: vscode.TextEditor, _edit: vscode.TextEditorEdit, ..._args: any[]) => {
-			cookCmdShortcutsOrFile(true, textEditor.document.uri.fsPath, false, false);
+			cookCmdShortcutsOrFile(true, textEditor.document.uri.fsPath, false, false, undefined, false, true);
 			cookCmdShortcutsOrFile(true, textEditor.document.uri.fsPath, false, true, undefined, false, true);
 			if (!isNullOrEmpty(getDefaultRootFolder())) {
 				cookCmdShortcutsOrFile(true, textEditor.document.uri.fsPath, true, false, undefined, false, false, false, true);
@@ -224,7 +224,7 @@ export function registerExtension(context: vscode.ExtensionContext) {
 
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('msr.cookCmdAliasDumpWithOthersToFilesByProject',
 		(textEditor: vscode.TextEditor, _edit: vscode.TextEditorEdit, ..._args: any[]) => {
-			cookCmdShortcutsOrFile(true, textEditor.document.uri.fsPath, true, false);
+			cookCmdShortcutsOrFile(true, textEditor.document.uri.fsPath, true, false, undefined, false, true);
 			cookCmdShortcutsOrFile(true, textEditor.document.uri.fsPath, true, true, undefined, false, true);
 			if (!isNullOrEmpty(getDefaultRootFolder())) {
 				cookCmdShortcutsOrFile(true, textEditor.document.uri.fsPath, true, false, undefined, false, false, false, true);